requirements.txt
onnx==1.14.0
psutil==5.9.5
pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
protobuf==4.23.2
gradio==3.40.0
tensorflow==2.15.0
customtkinter==5.2.0
tkinterdnd2==0.3.0
opencv-python==4.8.0.74
insightface==0.7.3
numpy==1.24.3
opennsfw2==0.9.2
tqdm==4.65.0
gfpgan==1.3.8
onnxruntime
insightface
